Rename EasyTier to SDWAN: Complete rebranding of luci-app-easytier to luci-app-sdwan
