+++ .github/workflows/build.yml
@@ -29,33 +29,33 @@ jobs:
       matrix:
         arch:
           - aarch64_cortex-a53
-          - aarch64_cortex-a72
-          - aarch64_generic
-          - arm_arm1176jzf-s_vfp
-          - arm_arm926ej-s
-          - arm_cortex-a15_neon-vfpv4
-          - arm_cortex-a5_vfpv4
-          - arm_cortex-a7
-          - arm_cortex-a7_neon-vfpv4
-          - arm_cortex-a8_vfpv3
-          - arm_cortex-a9
-          - arm_cortex-a9_neon
-          - arm_cortex-a9_vfpv3-d16
-          - arm_fa526
+          #- aarch64_cortex-a72
+          #- aarch64_generic
+          #- arm_arm1176jzf-s_vfp
+          #- arm_arm926ej-s
+          #- arm_cortex-a15_neon-vfpv4
+          #- arm_cortex-a5_vfpv4
+          #- arm_cortex-a7
+          #- arm_cortex-a7_neon-vfpv4
+          #- arm_cortex-a8_vfpv3
+          #- arm_cortex-a9
+          #- arm_cortex-a9_neon
+          #- arm_cortex-a9_vfpv3-d16
+          #- arm_fa526
           #- arm_mpcore
-          - arm_xscale
-          - mips_24kc
+          #- arm_xscale
+          #- mips_24kc
           #- mips_4kec
-          - mips_mips32
-          - mipsel_24kc
-          - mipsel_24kc_24kf
-          - mipsel_74kc
-          - mipsel_mips32
+          #- mips_mips32
+          #- mipsel_24kc
+          #- mipsel_24kc_24kf
+          #- mipsel_74kc
+          #- mipsel_mips32
           - x86_64
           
         sdk:
           - 22.03.7
-          - SNAPSHOT
+          #- SNAPSHOT
 
     steps:
       - uses: actions/checkout@v5
