--- luci-app-easytier/luasrc/model/cbi/easytierweb_log.lua
@@ -1,5 +1,0 @@
-f = SimpleForm("easytier")
-f.reset = false
-f.submit = false
-f:append(Template("easytier/easytierweb_log"))
-return f
