--- luci-app-easytier/po/zh-cn
@@ -1,1 +1,0 @@
-./zh_Hans
