--- luci-app-easytier/root/etc/config/easytier
@@ -1,7 +1,0 @@
-
-config easytier
-	option enabled '0'
-	option easytierbin '/usr/bin/easytier-core'
-	option auto_config_interface '1'
-	option auto_config_firewall '1'
-	option interface_netmask '255.0.0.0'
