--- luci-app-easytier/root/etc/uci-defaults/luci-easytier
@@ -1,13 +1,0 @@
-#!/bin/sh
-
-uci -q batch <<-EOF >/dev/null
-	delete ucitrack.@easytier[-1]
-	add ucitrack easytier
-	set ucitrack.@easytier[-1].init=easytier
-	commit ucitrack
-EOF
-
-chmod +x /etc/init.d/easytier
-/etc/init.d/easytier enable
-rm -rf /tmp/luci-indexcache /tmp/luci-modulecache
-exit 0
