--- luci-app-easytier/root/usr/share/rpcd/acl.d/luci-app-easytier.json
@@ -1,11 +1,0 @@
-{
-  "luci-app-easytier": {
-    "description": "Grant UCI access for luci-app-easytier",
-    "read": {
-      "uci": [ "easytier" ]
-    },
-    "write": {
-      "uci": [ "easytier" ]
-    }
-  }
-}
