+++ luci-app-sdwan/luasrc/model/cbi/sdwan_log.lua
@@ -1,0 +1,5 @@
+f = SimpleForm("sdwan")
+f.reset = false
+f.submit = false
+f:append(Template("sdwan/sdwan_log"))
+return f
