+++ luci-app-sdwan/luasrc/model/cbi/sdwanweb_log.lua
@@ -1,0 +1,5 @@
+f = SimpleForm("sdwan")
+f.reset = false
+f.submit = false
+f:append(Template("sdwan/sdwanweb_log"))
+return f
