+++ luci-app-sdwan/luasrc/view/sdwan/other_upload.htm
