+++ luci-app-sdwan/luasrc/view/sdwan/sdwanweb_log.htm
@@ -1,0 +1,49 @@
+<%
+local dsp = require "luci.dispatcher"
+-%>
+
+<script type="text/javascript">
+	//<![CDATA[
+	function clearlog(btn) {
+		XHR.get('<%=dsp.build_url("admin/vpn/sdwan/clear_wlog")%>', null,
+			function(x, data) {
+				if(x && x.status == 200) {
+					var log_textarea = document.getElementById('log_textarea');
+					log_textarea.innerHTML = "";
+					log_textarea.scrollTop = log_textarea.scrollHeight;
+				}
+			}
+		);
+	}
+	function downloadLog() {
+		var log_textarea = document.getElementById('log_textarea');
+		var logContent = log_textarea.value || log_textarea.innerText || log_textarea.innerHTML;
+		var blob = new Blob([logContent], {type: 'text/plain;charset=utf-8'});
+		var url = window.URL.createObjectURL(blob);
+		var a = document.createElement('a');
+		a.href = url;
+		a.download = 'sdwan-web_log.txt';  
+		document.body.appendChild(a);
+		a.click();
+		setTimeout(function() {
+			document.body.removeChild(a);
+			window.URL.revokeObjectURL(url);
+		}, 0);
+	}
+	XHR.poll(2, '<%=dsp.build_url("admin/vpn/sdwan/get_wlog")%>', null,
+		function(x, data) {
+			if(x && x.status == 200 && document.getElementById("checkbox1").checked == true) {
+				var log_textarea = document.getElementById('log_textarea');
+				log_textarea.innerHTML = x.responseText;
+				log_textarea.scrollTop = log_textarea.scrollHeight;
+			}
+		}
+	);
+	//]]>
+</script>
+<fieldset class="cbi-section" id="_log_fieldset">
+	<input type="checkbox" id="checkbox1" style="vertical-align:middle;height: auto;"checked><%:Auto Refresh%></input>
+	<input class="cbi-button cbi-input-remove" type="button" onclick="clearlog()" value="<%:Clear Logs%>" />
+	<input class="cbi-button cbi-input-apply" type="button" onclick="downloadLog()" value="<%:Download Log%>" />
+	<textarea id="log_textarea" class="cbi-input-textarea" style="width: 100%;margin-top: 10px;" data-update="change" rows="30" wrap="off" readonly="readonly"></textarea>
+</fieldset>
