+++ luci-app-sdwan/po/zh-cn
@@ -1,0 +1,1 @@
+./zh_Hans
