+++ luci-app-sdwan/root/etc/config/sdwan
@@ -1,0 +1,7 @@
+
+config sdwan
+	option enabled '0'
+	option sdwanbin '/usr/bin/sdwan-core'
+	option auto_config_interface '1'
+	option auto_config_firewall '1'
+	option interface_netmask '255.0.0.0'
