+++ luci-app-sdwan/root/etc/uci-defaults/luci-sdwan
@@ -1,0 +1,1 @@
+exit 0
