+++ luci-app-sdwan/root/usr/share/rpcd/acl.d/luci-app-sdwan.json
@@ -1,0 +1,11 @@
+{
+  "luci-app-sdwan": {
+    "description": "Grant UCI access for luci-app-sdwan",
+    "read": {
+      "uci": [ "sdwan" ]
+    },
+    "write": {
+      "uci": [ "sdwan" ]
+    }
+  }
+}
