+++ luci-app-sdwan/root/usr/share/sdwan/download.sh
@@ -1,0 +1,1 @@
+}
