+++ luci-app-sdwan/root/usr/share/sdwan/firewall.sh
@@ -1,0 +1,1 @@
+}
