+++ version.mk
@@ -1,5 +1,5 @@
-# EasyTier Version Configuration
-# This file is the single source of truth for the EasyTier version
-# Used by: easytier/Makefile, luci-app-easytier/Makefile, init.d/easytier
+# SDWAN Version Configuration
+# This file is the single source of truth for the SDWAN version
+# Used by: sdwan/Makefile, luci-app-sdwan/Makefile, init.d/sdwan
 
-EASYTIER_VERSION=2.5.0
+SDWAN_VERSION=2.5.0
